requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.16.0
chardet==3.0.4
cryptography==2.3
idna==2.6
keyring==10.6.0
keyrings.alt==3.0
numpy==1.19.3
opencv-python==4.5.1.48
Pillow==6.2.2
pycairo==1.26.0
pycparser==2.22
pycrypto==2.6.1
pytesseract==0.3.0
pyxdg==0.26
requests==2.22.0
SecretStorage==2.3.1
six==1.11.0
urllib3==1.25.7
